chore(pom): agrega dependencia AssertJ y corrige test ambiguo
Agrega assertj-core 3.26.3 como dependencia de test para las pruebas con AssertJ. Corrige llamada ambigua a MenuFactory.crearMenu(null, platos) - se hace cast explicito a Integer.

+++ src/test/java/co/edu/udec/guarderia/domain/factory/MenuFactoryTest.java
@@ -28,7 +28,7 @@ class MenuFactoryTest {
         var platos = List.of(new Plato(new NombrePlato("Sopa"),
                 List.of(new NombreIngrediente("Agua")), BigDecimal.ONE));
 
-        assertThatThrownBy(() -> MenuFactory.crearMenu(null, platos))
+        assertThatThrownBy(() -> MenuFactory.crearMenu((Integer) null, platos))
                 .isInstanceOf(RuntimeException.class)
                 .hasMessageContaining("ID del menu");
     }
